requirements.txt
annotated-types==0.7.0
anyio==4.10.0
beautifulsoup4==4.12.2
boto3==1.40.30
botocore==1.40.30
certifi==2025.8.3
charset-normalizer==3.4.3
distro==1.9.0
flask==3.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.75
langchain-openai==0.3.32
langsmith==0.4.21
lxml==4.9.3
openai==1.102.0
orjson==3.11.3
packaging==25.0
py4j==0.10.9.9
pydantic==2.11.7
pydantic_core==2.33.2
pyspark==4.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.2
regex==2025.8.29
requests==2.32.5
requests-toolbelt==1.0.0
s3transfer==0.14.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
tenacity==9.1.2
tiktoken==0.11.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
wheel==0.45.1
zstandard==0.24.0
